scripts/requirements.txt
torch
nni
tensorboard
onnxruntime-gpu
coloredlogs
onnx
sympy
transformers
pyarrow
ray
elasticsearch